requirements/dev.txt
-r base.txt
django-extensions==1.2.0
ipdb==0.7
ipython==0.13.2
coverage==3.6
rednose==0.4.1
django-nose==1.2
django-debug-toolbar==0.10.2
dj-static
